html/canvas/offscreen/canvas-host/2d.canvas.host.size.attributes.reflect.setcontent.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.size.attributes.reflect.setcontent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.attributes.reflect.setcontent</h1>
<p class="desc">Setting content attributes updates IDL and content attributes</p>


<script>
var t = async_test("Setting content attributes updates IDL and content attributes");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.setAttribute('width', '120');
  canvas.setAttribute('height', '60');
  _assertSame(canvas.getAttribute('width'), '120', "canvas.getAttribute('width')", "'120'");
  _assertSame(canvas.getAttribute('height'), '60', "canvas.getAttribute('height')", "'60'");
  _assertSame(canvas.width, 120, "canvas.width", "120");
  _assertSame(canvas.height, 60, "canvas.height", "60");
  t.done();

});
</script>
